Create README.md
Welcome to the Data Science (ML and DL) Projects repository! 🎉

This repository is a curated collection of projects that demonstrate the power of Machine Learning (ML) and Deep Learning (DL) in solving real-world problems. Whether you're a beginner looking to explore data science or an experienced professional, this repo has something for everyone.